Fix: Removed unnecessary comments + deepMergeTranslations tests in wrong file
